refactored mock plugin setup
