+++ calendar/utils/kvstore/plugin_store_test.go
@@ -4,70 +4,24 @@ import (
 	"testing"
 	"time"
 
-	"github.com/mattermost/mattermost/server/public/model"
-	"github.com/mattermost/mattermost/server/public/plugin"
-	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
+
+	"github.com/mattermost/mattermost/server/public/model"
+
+	"github.com/mattermost/mattermost-plugin-mscalendar/calendar/testutil"
 )
-
-type MockPluginAPI struct {
-	plugin.API
-	mock.Mock
-}
-
-func (m *MockPluginAPI) KVGet(key string) ([]byte, *model.AppError) {
-	args := m.Called(key)
-	data, _ := args.Get(0).([]byte)
-	if err := args.Get(1); err != nil {
-		return nil, err.(*model.AppError)
-	}
-	return data, nil
-}
-
-func (m *MockPluginAPI) KVSet(key string, data []byte) *model.AppError {
-	args := m.Called(key, data)
-	if err := args.Get(0); err != nil {
-		return err.(*model.AppError)
-	}
-	return nil
-}
-
-func (m *MockPluginAPI) KVSetWithExpiry(key string, data []byte, ttl int64) *model.AppError {
-	args := m.Called(key, data, ttl)
-	if err := args.Get(0); err != nil {
-		return err.(*model.AppError)
-	}
-	return nil
-}
-
-func (m *MockPluginAPI) KVSetWithOptions(key string, value []byte, options model.PluginKVSetOptions) (bool, *model.AppError) {
-	args := m.Called(key, value, options)
-	success := args.Bool(0)
-	if err := args.Get(1); err != nil {
-		return success, err.(*model.AppError)
-	}
-	return success, nil
-}
-
-func (m *MockPluginAPI) KVDelete(key string) *model.AppError {
-	args := m.Called(key)
-	if err := args.Get(0); err != nil {
-		return err.(*model.AppError)
-	}
-	return nil
-}
 
 func TestLoad(t *testing.T) {
 	tests := []struct {
 		name       string
 		key        string
-		setup      func(*MockPluginAPI)
+		setup      func(*testutil.MockPluginAPI)
 		assertions func(*testing.T, []byte, error)
 	}{
 		{
 			name: "Error during KVGet",
 			key:  "error-key",
-			setup: func(mockAPI *MockPluginAPI) {
+			setup: func(mockAPI *testutil.MockPluginAPI) {
 				mockAPI.On("KVGet", "error-key").Return(nil, &model.AppError{Message: "KVGet failed"})
 			},
 			assertions: func(t *testing.T, data []byte, err error) {
@@ -78,7 +32,7 @@ func TestLoad(t *testing.T) {
 		{
 			name: "Key not found",
 			key:  "missing-key",
-			setup: func(mockAPI *MockPluginAPI) {
+			setup: func(mockAPI *testutil.MockPluginAPI) {
 				mockAPI.On("KVGet", "missing-key").Return(nil, nil)
 			},
 			assertions: func(t *testing.T, data []byte, err error) {
@@ -89,7 +43,7 @@ func TestLoad(t *testing.T) {
 		{
 			name: "Load successfully",
 			key:  "test-key",
-			setup: func(mockAPI *MockPluginAPI) {
+			setup: func(mockAPI *testutil.MockPluginAPI) {
 				mockAPI.On("KVGet", "test-key").Return([]byte("test-value"), nil)
 			},
 			assertions: func(t *testing.T, data []byte, err error) {
@@ -101,7 +55,7 @@ func TestLoad(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			mockAPI := &MockPluginAPI{}
+			mockAPI := &testutil.MockPluginAPI{}
 			store := NewPluginStore(mockAPI)
 			tt.setup(mockAPI)
 
@@ -118,13 +72,13 @@ func TestStore(t *testing.T) {
 	tests := []struct {
 		name       string
 		expiryTime int
-		setup      func(*MockPluginAPI)
+		setup      func(*testutil.MockPluginAPI)
 		assertions func(*testing.T, error)
 	}{
 		{
 			name:       "Error during KVSet with TTL",
 			expiryTime: 60,
-			setup: func(mockAPI *MockPluginAPI) {
+			setup: func(mockAPI *testutil.MockPluginAPI) {
 				mockAPI.On("KVSetWithExpiry", "mockKey", []byte("mockValue"), int64(60)).Return(&model.AppError{Message: "KVSet failed"})
 			},
 			assertions: func(t *testing.T, err error) {
@@ -134,7 +88,7 @@ func TestStore(t *testing.T) {
 		{
 			name:       "Error during KVSet without TTL",
 			expiryTime: 0,
-			setup: func(mockAPI *MockPluginAPI) {
+			setup: func(mockAPI *testutil.MockPluginAPI) {
 				mockAPI.On("KVSet", "mockKey", []byte("mockValue")).Return(&model.AppError{Message: "KVSet failed"})
 			},
 			assertions: func(t *testing.T, err error) {
@@ -144,7 +98,7 @@ func TestStore(t *testing.T) {
 		{
 			name:       "Store with TTL successfully",
 			expiryTime: 60,
-			setup: func(mockAPI *MockPluginAPI) {
+			setup: func(mockAPI *testutil.MockPluginAPI) {
 				mockAPI.On("KVSetWithExpiry", "mockKey", []byte("mockValue"), int64(60)).Return(nil)
 			},
 			assertions: func(t *testing.T, err error) {
@@ -154,7 +108,7 @@ func TestStore(t *testing.T) {
 		{
 			name:       "Store without TTL successfully",
 			expiryTime: 0,
-			setup: func(mockAPI *MockPluginAPI) {
+			setup: func(mockAPI *testutil.MockPluginAPI) {
 				mockAPI.On("KVSet", "mockKey", []byte("mockValue")).Return(nil)
 			},
 			assertions: func(t *testing.T, err error) {
@@ -165,7 +119,7 @@ func TestStore(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			mockAPI := &MockPluginAPI{}
+			mockAPI := &testutil.MockPluginAPI{}
 			store := NewPluginStoreWithExpiry(mockAPI, time.Duration(tt.expiryTime)*time.Second)
 			tt.setup(mockAPI)
 
@@ -181,12 +135,12 @@ func TestStore(t *testing.T) {
 func TestStoreTTL(t *testing.T) {
 	tests := []struct {
 		name       string
-		setup      func(*MockPluginAPI)
+		setup      func(*testutil.MockPluginAPI)
 		assertions func(*testing.T, error)
 	}{
 		{
 			name: "Error during storing with TTL",
-			setup: func(mockAPI *MockPluginAPI) {
+			setup: func(mockAPI *testutil.MockPluginAPI) {
 				mockAPI.On("KVSetWithExpiry", "mockKey", []byte("mockValue"), int64(60)).Return(&model.AppError{Message: "KVSet failed"})
 			},
 			assertions: func(t *testing.T, err error) {
@@ -195,7 +149,7 @@ func TestStoreTTL(t *testing.T) {
 		},
 		{
 			name: "Store with TTL successfully",
-			setup: func(mockAPI *MockPluginAPI) {
+			setup: func(mockAPI *testutil.MockPluginAPI) {
 				mockAPI.On("KVSetWithExpiry", "mockKey", []byte("mockValue"), int64(60)).Return(nil)
 			},
 			assertions: func(t *testing.T, err error) {
@@ -205,7 +159,7 @@ func TestStoreTTL(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			mockAPI := &MockPluginAPI{}
+			mockAPI := &testutil.MockPluginAPI{}
 			store := NewPluginStoreWithExpiry(mockAPI, 60*time.Second)
 			tt.setup(mockAPI)
 
@@ -223,7 +177,7 @@ func TestStoreWithOptions(t *testing.T) {
 		name       string
 		expiryTime int64
 		opts       model.PluginKVSetOptions
-		setup      func(*MockPluginAPI)
+		setup      func(*testutil.MockPluginAPI)
 		assertions func(*testing.T, bool, error)
 	}{
 		{
@@ -232,7 +186,7 @@ func TestStoreWithOptions(t *testing.T) {
 			opts: model.PluginKVSetOptions{
 				ExpireInSeconds: 30,
 			},
-			setup: func(mockAPI *MockPluginAPI) {
+			setup: func(mockAPI *testutil.MockPluginAPI) {
 				mockAPI.On("KVSetWithOptions", "mockKey", []byte("mockValue"), model.PluginKVSetOptions{ExpireInSeconds: 30}).Return(false, &model.AppError{Message: "KVSet failed"})
 			},
 			assertions: func(t *testing.T, success bool, err error) {
@@ -244,7 +198,7 @@ func TestStoreWithOptions(t *testing.T) {
 			name:       "Use default TTL when opts.ExpireInSeconds is 0",
 			expiryTime: 60,
 			opts:       model.PluginKVSetOptions{},
-			setup: func(mockAPI *MockPluginAPI) {
+			setup: func(mockAPI *testutil.MockPluginAPI) {
 				mockAPI.On("KVSetWithOptions", "mockKey", []byte("mockValue"), model.PluginKVSetOptions{ExpireInSeconds: 60}).Return(true, nil)
 			},
 			assertions: func(t *testing.T, success bool, err error) {
@@ -258,7 +212,7 @@ func TestStoreWithOptions(t *testing.T) {
 			opts: model.PluginKVSetOptions{
 				ExpireInSeconds: 30,
 			},
-			setup: func(mockAPI *MockPluginAPI) {
+			setup: func(mockAPI *testutil.MockPluginAPI) {
 				mockAPI.On("KVSetWithOptions", "mockKey", []byte("mockValue"), model.PluginKVSetOptions{ExpireInSeconds: 30}).Return(true, nil)
 			},
 			assertions: func(t *testing.T, success bool, err error) {
@@ -269,7 +223,7 @@ func TestStoreWithOptions(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			mockAPI := &MockPluginAPI{}
+			mockAPI := &testutil.MockPluginAPI{}
 			store := NewPluginStoreWithExpiry(mockAPI, time.Duration(tt.expiryTime)*time.Second)
 			tt.setup(mockAPI)
 
@@ -285,12 +239,12 @@ func TestStoreWithOptions(t *testing.T) {
 func TestDelete(t *testing.T) {
 	tests := []struct {
 		name       string
-		setup      func(*MockPluginAPI)
+		setup      func(*testutil.MockPluginAPI)
 		assertions func(*testing.T, error)
 	}{
 		{
 			name: "Error during KVDelete",
-			setup: func(mockAPI *MockPluginAPI) {
+			setup: func(mockAPI *testutil.MockPluginAPI) {
 				mockAPI.On("KVDelete", "mockKey").Return(&model.AppError{Message: "KVDelete failed"})
 			},
 			assertions: func(t *testing.T, err error) {
@@ -299,7 +253,7 @@ func TestDelete(t *testing.T) {
 		},
 		{
 			name: "Delete successfully",
-			setup: func(mockAPI *MockPluginAPI) {
+			setup: func(mockAPI *testutil.MockPluginAPI) {
 				mockAPI.On("KVDelete", "mockKey").Return(nil)
 			},
 			assertions: func(t *testing.T, err error) {
@@ -309,7 +263,7 @@ func TestDelete(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			mockAPI := &MockPluginAPI{}
+			mockAPI := &testutil.MockPluginAPI{}
 			store := NewPluginStore(mockAPI)
 			tt.setup(mockAPI)
